Move extant tracing API into new pkg
